feat(rpc): Create celestia-rpc crate and add integration tests

+++ types/src/error.rs
@@ -32,7 +32,7 @@ pub enum Error {
     MissingProof,
 
     #[error("Unsupported share version: {0}")]
-    UnsupportedShareVersion(u32),
+    UnsupportedShareVersion(u8),
 
     #[error("Invalid share size: {0}")]
     InvalidShareSize(usize),
